requirements.txt
google-api-python-client==1.6.6
Flask==0.12.0
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
Werkzeug==0.14.1
click==6.7
python-dotenv==0.8.2
Flask-Script==2.0.6
gspread==3.0.0